+++ express/blocks/pricing-summary/pricing-summary.js
@@ -11,82 +11,7 @@
  */
 
 import { createTag } from '../../scripts/scripts.js';
-import { getOffer, buildUrl } from '../../scripts/utils/pricing.js';
-
-async function fetchPlan(planUrl) {
-  if (!window.pricingPlans) {
-    window.pricingPlans = {};
-  }
-
-  let plan = window.pricingPlans[planUrl];
-
-  if (!plan) {
-    plan = {};
-    const link = new URL(planUrl);
-    const params = link.searchParams;
-
-    plan.url = planUrl;
-    plan.country = 'us';
-    plan.language = 'en';
-    plan.price = '9.99';
-    plan.currency = 'US';
-    plan.symbol = '$';
-
-    // TODO: Remove '/sp/ once confirmed with stakeholders
-    const allowedHosts = ['new.express.adobe.com', 'express.adobe.com', 'adobesparkpost.app.link'];
-    const { host } = new URL(planUrl);
-    if (allowedHosts.includes(host) || planUrl.includes('/sp/')) {
-      plan.offerId = 'FREE0';
-      plan.frequency = 'monthly';
-      plan.name = 'Free';
-      plan.stringId = 'free-trial';
-    } else {
-      plan.offerId = params.get('items[0][id]');
-      plan.frequency = null;
-      plan.name = 'Premium';
-      plan.stringId = '3-month-trial';
-    }
-
-    if (plan.offerId === '70C6FDFC57461D5E449597CC8F327CF1' || plan.offerId === 'CFB1B7F391F77D02FE858C43C4A5C64F') {
-      plan.frequency = 'Monthly';
-    } else if (plan.offerId === 'E963185C442F0C5EEB3AE4F4AAB52C24' || plan.offerId === 'BADDACAB87D148A48539B303F3C5FA92') {
-      plan.frequency = 'Annual';
-    } else {
-      plan.frequency = null;
-    }
-
-    const countryOverride = new URLSearchParams(window.location.search).get('country');
-    const offer = await getOffer(plan.offerId, countryOverride);
-
-    if (offer) {
-      plan.currency = offer.currency;
-      plan.price = offer.unitPrice;
-      plan.basePrice = offer.basePrice;
-      plan.country = offer.country;
-      plan.vatInfo = offer.vatInfo;
-      plan.language = offer.lang;
-      plan.rawPrice = offer.unitPriceCurrencyFormatted.match(/[\d\s,.+]+/g);
-      plan.prefix = offer.prefix ?? '';
-      plan.suffix = offer.suffix ?? '';
-      plan.formatted = offer.unitPriceCurrencyFormatted.replace(
-        plan.rawPrice[0],
-        `<strong>${plan.prefix}${plan.rawPrice[0]}</strong>`,
-      );
-
-      if (offer.basePriceCurrencyFormatted) {
-        plan.rawBasePrice = offer.basePriceCurrencyFormatted.match(/[\d\s,.+]+/g);
-        plan.formattedBP = offer.basePriceCurrencyFormatted.replace(
-          plan.rawBasePrice[0],
-          `<strong>${plan.prefix}${plan.rawBasePrice[0]}</strong>`,
-        );
-      }
-    }
-
-    window.pricingPlans[planUrl] = plan;
-  }
-
-  return plan;
-}
+import { fetchPlan, buildUrl } from '../../scripts/utils/pricing.js';
 
 function handleHeader(column) {
   column.classList.add('pricing-column');
